rename caddy/ntlm to ntlm
